requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
django-cors-headers==4.7.0
drf-yasg==1.21.10
